requirements.txt
asgiref==3.7.2
Django==4.2.3
django-watson==1.6.3
psycopg==3.1.9
sqlparse==0.4.4
typing_extensions==4.7.1
